# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"ML Pipeline Thing API","description":"ML Pipeline Thing API","version":"0.1.0"},"paths":{"/health/":{"get":{"tags":["Health"],"summary":"Health Check","operationId":"health_check_health__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthCheck"}}}}}}},"/v1/pipelines/":{"post":{"tags":["Pipelines"],"summary":"Create Pipeline","description":"Register a new ML pipeline with its stages","operationId":"create_pipeline_v1_pipelines__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PipelineCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PipelineWithStages"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Pipelines"],"summary":"List Pipelines","description":"List all registered pipelines with pagination","operationId":"list_pipelines_v1_pipelines__get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginationResponse_PipelineResponse_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/pipelines/{pipeline_id}":{"get":{"tags":["Pipelines"],"summary":"Get Pipeline","description":"Get detailed information about a specific pipeline","operationId":"get_pipeline_v1_pipelines__pipeline_id__get","parameters":[{"name":"pipeline_id","in":"path","required":true,"schema":{"type":"string","title":"Pipeline Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PipelineWithStages"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Pipelines"],"summary":"Delete Pipeline","description":"Delete a pipeline and all its associated data","operationId":"delete_pipeline_v1_pipelines__pipeline_id__delete","parameters":[{"name":"pipeline_id","in":"path","required":true,"schema":{"type":"string","title":"Pipeline Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/pipelines/{pipeline_id}/trigger_run":{"post":{"tags":["Pipeline Runs"],"summary":"Trigger Pipeline Run","description":"Start a new pipeline run","operationId":"trigger_pipeline_run_v1_pipelines__pipeline_id__trigger_run_post","parameters":[{"name":"pipeline_id","in":"path","required":true,"schema":{"type":"string","title":"Pipeline Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TriggerRunRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PipelineRunResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/pipelines/{pipeline_id}/runs":{"get":{"tags":["Pipeline Runs"],"summary":"List Pipeline Runs","description":"List all runs for a specific pipeline with pagination","operationId":"list_pipeline_runs_v1_pipelines__pipeline_id__runs_get","parameters":[{"name":"pipeline_id","in":"path","required":true,"schema":{"type":"string","title":"Pipeline Id"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginationResponse_PipelineRunResponse_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/pipelines/{pipeline_id}/runs/{run_id}":{"get":{"tags":["Pipeline Runs"],"summary":"Get Pipeline Run","description":"Get detailed information about a specific pipeline run","operationId":"get_pipeline_run_v1_pipelines__pipeline_id__runs__run_id__get","parameters":[{"name":"pipeline_id","in":"path","required":true,"schema":{"type":"string","title":"Pipeline Id"}},{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PipelineRunWithStages"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Pipeline Runs"],"summary":"Cancel Pipeline Run","description":"Cancel a running pipeline","operationId":"cancel_pipeline_run_v1_pipelines__pipeline_id__runs__run_id__delete","parameters":[{"name":"pipeline_id","in":"path","required":true,"schema":{"type":"string","title":"Pipeline Id"}},{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthCheck":{"properties":{"status":{"type":"string","title":"Status","default":"healthy"},"version":{"type":"string","title":"Version"},"database":{"type":"string","title":"Database","default":"connected"},"uptime":{"type":"number","title":"Uptime"},"environment":{"type":"string","title":"Environment"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"memoryUsage":{"$ref":"#/components/schemas/MemoryUsage"}},"type":"object","required":["version","uptime","environment","timestamp","memoryUsage"],"title":"HealthCheck","description":"Health check response model"},"MemoryUsage":{"properties":{"used":{"type":"string","title":"Used"},"available":{"type":"string","title":"Available"},"percent":{"type":"string","title":"Percent"}},"type":"object","required":["used","available","percent"],"title":"MemoryUsage","description":"Memory usage model"},"PaginationResponse_PipelineResponse_":{"properties":{"items":{"items":{"$ref":"#/components/schemas/PipelineResponse"},"type":"array","title":"Items","description":"List of items for the current page"},"total":{"type":"integer","title":"Total","description":"Total number of items across all pages"},"page":{"type":"integer","title":"Page","description":"Current page number (1-based)"},"size":{"type":"integer","title":"Size","description":"Number of items per page"},"pages":{"type":"integer","title":"Pages","description":"Total number of pages"},"hasNext":{"type":"boolean","title":"Hasnext","description":"Whether there is a next page"},"hasPrev":{"type":"boolean","title":"Hasprev","description":"Whether there is a previous page"}},"type":"object","required":["items","total","page","size","pages","hasNext","hasPrev"],"title":"PaginationResponse[PipelineResponse]"},"PaginationResponse_PipelineRunResponse_":{"properties":{"items":{"items":{"$ref":"#/components/schemas/PipelineRunResponse"},"type":"array","title":"Items","description":"List of items for the current page"},"total":{"type":"integer","title":"Total","description":"Total number of items across all pages"},"page":{"type":"integer","title":"Page","description":"Current page number (1-based)"},"size":{"type":"integer","title":"Size","description":"Number of items per page"},"pages":{"type":"integer","title":"Pages","description":"Total number of pages"},"hasNext":{"type":"boolean","title":"Hasnext","description":"Whether there is a next page"},"hasPrev":{"type":"boolean","title":"Hasprev","description":"Whether there is a previous page"}},"type":"object","required":["items","total","page","size","pages","hasNext","hasPrev"],"title":"PaginationResponse[PipelineRunResponse]"},"PipelineCreate":{"properties":{"name":{"type":"string","maxLength":255,"minLength":1,"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Config"},"stages":{"items":{"$ref":"#/components/schemas/PipelineStageCreate"},"type":"array","minItems":1,"title":"Stages"}},"type":"object","required":["name","stages"],"title":"PipelineCreate"},"PipelineResponse":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"status":{"type":"string","title":"Status"},"config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Config"},"startedAt":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Startedat"},"completedAt":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completedat"},"executionTime":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Executiontime"},"memoryUsage":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Memoryusage"},"cpuUsage":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Cpuusage"},"createdAt":{"type":"string","format":"date-time","title":"Createdat"},"updatedAt":{"type":"string","format":"date-time","title":"Updatedat"}},"type":"object","required":["id","name","status","createdAt","updatedAt"],"title":"PipelineResponse"},"PipelineRunResponse":{"properties":{"id":{"type":"string","title":"Id"},"pipelineId":{"type":"string","title":"Pipelineid"},"status":{"$ref":"#/components/schemas/RunStatus"},"triggerType":{"$ref":"#/components/schemas/TriggerType"},"triggeredBy":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Triggeredby"},"startedAt":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Startedat"},"completedAt":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completedat"},"executionTime":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Executiontime"},"runConfig":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Runconfig"},"environment":{"type":"string","title":"Environment"},"maxMemoryUsage":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Maxmemoryusage"},"maxCpuUsage":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Maxcpuusage"},"successCount":{"type":"integer","title":"Successcount","default":0},"failedCount":{"type":"integer","title":"Failedcount","default":0},"errorMessage":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Errormessage"},"outputData":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Outputdata"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"},"createdAt":{"type":"string","format":"date-time","title":"Createdat"},"updatedAt":{"type":"string","format":"date-time","title":"Updatedat"}},"type":"object","required":["id","pipelineId","status","triggerType","environment","createdAt","updatedAt"],"title":"PipelineRunResponse"},"PipelineRunWithStages":{"properties":{"id":{"type":"string","title":"Id"},"pipelineId":{"type":"string","title":"Pipelineid"},"status":{"$ref":"#/components/schemas/RunStatus"},"triggerType":{"$ref":"#/components/schemas/TriggerType"},"triggeredBy":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Triggeredby"},"startedAt":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Startedat"},"completedAt":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completedat"},"executionTime":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Executiontime"},"runConfig":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Runconfig"},"environment":{"type":"string","title":"Environment"},"maxMemoryUsage":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Maxmemoryusage"},"maxCpuUsage":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Maxcpuusage"},"successCount":{"type":"integer","title":"Successcount","default":0},"failedCount":{"type":"integer","title":"Failedcount","default":0},"errorMessage":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Errormessage"},"outputData":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Outputdata"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"},"createdAt":{"type":"string","format":"date-time","title":"Createdat"},"updatedAt":{"type":"string","format":"date-time","title":"Updatedat"},"stageRuns":{"items":{"$ref":"#/components/schemas/StageRunResponse"},"type":"array","title":"Stageruns","default":[]}},"type":"object","required":["id","pipelineId","status","triggerType","environment","createdAt","updatedAt"],"title":"PipelineRunWithStages"},"PipelineStageCreate":{"properties":{"name":{"type":"string","maxLength":255,"minLength":1,"title":"Name"},"stageType":{"$ref":"#/components/schemas/StageType"},"customName":{"anyOf":[{"type":"string","maxLength":255,"minLength":1},{"type":"null"}],"title":"Customname"},"order":{"type":"integer","minimum":0.0,"title":"Order"},"config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Config"},"dependencies":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Dependencies"}},"type":"object","required":["name","stageType","order"],"title":"PipelineStageCreate"},"PipelineStageResponse":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"stageType":{"$ref":"#/components/schemas/StageType"},"customName":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Customname"},"status":{"$ref":"#/components/schemas/StageStatus"},"order":{"type":"integer","title":"Order"},"config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Config"},"dependencies":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Dependencies"},"startedAt":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Startedat"},"completedAt":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completedat"},"executionTime":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Executiontime"},"outputPath":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Outputpath"},"metrics":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Metrics"},"createdAt":{"type":"string","format":"date-time","title":"Createdat"},"updatedAt":{"type":"string","format":"date-time","title":"Updatedat"}},"type":"object","required":["id","name","stageType","status","order","createdAt","updatedAt"],"title":"PipelineStageResponse"},"PipelineWithStages":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"status":{"type":"string","title":"Status"},"config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Config"},"startedAt":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Startedat"},"completedAt":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completedat"},"executionTime":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Executiontime"},"memoryUsage":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Memoryusage"},"cpuUsage":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Cpuusage"},"createdAt":{"type":"string","format":"date-time","title":"Createdat"},"updatedAt":{"type":"string","format":"date-time","title":"Updatedat"},"stages":{"items":{"$ref":"#/components/schemas/PipelineStageResponse"},"type":"array","title":"Stages","default":[]}},"type":"object","required":["id","name","status","createdAt","updatedAt"],"title":"PipelineWithStages"},"RunStatus":{"type":"string","enum":["PENDING","RUNNING","COMPLETED","FAILED","CANCELLED"],"title":"RunStatus"},"StageRunResponse":{"properties":{"id":{"type":"string","title":"Id"},"pipelineRunId":{"type":"string","title":"Pipelinerunid"},"stageId":{"type":"string","title":"Stageid"},"status":{"$ref":"#/components/schemas/RunStatus"},"attemptNumber":{"type":"integer","title":"Attemptnumber","default":1},"startedAt":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Startedat"},"completedAt":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completedat"},"executionTime":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Executiontime"},"memoryUsage":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Memoryusage"},"cpuUsage":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Cpuusage"},"outputData":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Outputdata"},"errorMessage":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Errormessage"},"logs":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Logs"},"createdAt":{"type":"string","format":"date-time","title":"Createdat"},"updatedAt":{"type":"string","format":"date-time","title":"Updatedat"}},"type":"object","required":["id","pipelineRunId","stageId","status","createdAt","updatedAt"],"title":"StageRunResponse"},"StageStatus":{"type":"string","enum":["PENDING","RUNNING","COMPLETED","FAILED","SKIPPED"],"title":"StageStatus"},"StageType":{"type":"string","enum":["DATA_INGESTION","DATA_VALIDATION","DATA_PREPROCESSING","FEATURE_ENGINEERING","DATA_SPLITTING","MODEL_TRAINING","MODEL_VALIDATION","MODEL_EVALUATION","MODEL_TESTING","MODEL_REGISTRATION","MODEL_DEPLOYMENT","MODEL_MONITORING","EXPLORATORY_DATA_ANALYSIS","HYPERPARAMETER_TUNING","MODEL_COMPARISON","ENVIRONMENT_SETUP","RESOURCE_PROVISIONING","CLEANUP","CUSTOM"],"title":"StageType"},"TriggerRunRequest":{"properties":{"runConfig":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Runconfig"},"environment":{"type":"string","pattern":"^(development|staging|production)$","title":"Environment","default":"development"},"triggeredBy":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Triggeredby"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","title":"TriggerRunRequest"},"TriggerType":{"type":"string","enum":["MANUAL","SCHEDULED","API","WEBHOOK"],"title":"TriggerType"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}